moved slider to own file
