+++ webxr/src/components/SliderComponent/models/knob.glb
